+++ migrations/Version20240911141618.php
@@ -1,0 +1,35 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DoctrineMigrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * Auto-generated Migration: Please modify to your needs!
+ */
+final class Version20240911141618 extends AbstractMigration
+{
+    public function getDescription(): string
+    {
+        return '';
+    }
+
+    public function up(Schema $schema): void
+    {
+        // this up() migration is auto-generated, please modify it to your needs
+        $this->addSql('ALTER TABLE person ADD primary_address_id INT DEFAULT NULL');
+        $this->addSql('ALTER TABLE person ADD CONSTRAINT FK_34DCD176CB134313 FOREIGN KEY (primary_address_id) REFERENCES address (id)');
+        $this->addSql('CREATE INDEX IDX_34DCD176CB134313 ON person (primary_address_id)');
+    }
+
+    public function down(Schema $schema): void
+    {
+        // this down() migration is auto-generated, please modify it to your needs
+        $this->addSql('ALTER TABLE person DROP FOREIGN KEY FK_34DCD176CB134313');
+        $this->addSql('DROP INDEX IDX_34DCD176CB134313 ON person');
+        $this->addSql('ALTER TABLE person DROP primary_address_id');
+    }
+}