+++ src/Controller/Admin/AbstractCrudController.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Controller\Admin;
+
+use EasyCorp\Bundle\EasyAdminBundle\Config\Crud;
+use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractCrudController as BaseAbstractCrudController;
+
+abstract class AbstractCrudController extends BaseAbstractCrudController
+{
+    public function configureCrud(Crud $crud): Crud
+    {
+        return $crud->showEntityActionsInlined();
+    }
+}