+++ src/Controller/Admin/DashboardController.php
@@ -1,0 +1,38 @@
+<?php
+
+namespace App\Controller\Admin;
+
+use App\Entity\Address;
+use App\Entity\Person;
+use EasyCorp\Bundle\EasyAdminBundle\Config\Dashboard;
+use EasyCorp\Bundle\EasyAdminBundle\Config\MenuItem;
+use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractDashboardController;
+use EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGenerator;
+use EasyCorp\Bundle\EasyAdminBundle\Router\AdminUrlGeneratorInterface;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Translation\TranslatableMessage;
+
+class DashboardController extends AbstractDashboardController
+{
+    #[Route('/admin', name: 'admin')]
+    public function index(): Response
+    {
+        /** @var AdminUrlGeneratorInterface $adminUrlGenerator */
+        $adminUrlGenerator = $this->container->get(AdminUrlGenerator::class);
+
+        return $this->redirect($adminUrlGenerator->setController(PersonCrudController::class)->generateUrl());
+    }
+
+    public function configureDashboard(): Dashboard
+    {
+        return Dashboard::new()
+            ->setTitle('App');
+    }
+
+    public function configureMenuItems(): iterable
+    {
+        yield MenuItem::linkToCrud(new TranslatableMessage('People'), 'fas fa-person', Person::class);
+        yield MenuItem::linkToCrud(new TranslatableMessage('Addresses'), 'fas fa-list', Address::class);
+    }
+}