+++ src/Controller/Admin/PersonCrudController.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace App\Controller\Admin;
+
+use App\Entity\Person;
+use EasyCorp\Bundle\EasyAdminBundle\Field\AssociationField;
+use EasyCorp\Bundle\EasyAdminBundle\Field\IdField;
+use EasyCorp\Bundle\EasyAdminBundle\Field\TextField;
+
+class PersonCrudController extends AbstractCrudController
+{
+    public static function getEntityFqcn(): string
+    {
+        return Person::class;
+    }
+
+    public function configureFields(string $pageName): iterable
+    {
+        yield IdField::new('id')->hideOnForm();
+        yield TextField::new('name');
+        yield AssociationField::new('primaryAddress');
+        yield AssociationField::new('address');
+    }
+}